refactor confirm to put the registry in a module
